move fadeEffect from label header to its own file
